+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * UX Consistency: Update HTML Page Title display when third pane (detail record) displays. Refs UIREC-319.
 * Add "Display to public" toggle on piece form. Refs UIREC-333.
+
+## [5.0.2](https://github.com/folio-org/ui-receiving/tree/v5.0.2) (2024-04-18)
+[Full Changelog](https://github.com/folio-org/ui-receiving/compare/v5.0.1...v5.0.2)
+
 * Add additional receiving filters to support reporting requirements. Refs UIREC-341.
 
 ## [5.0.1](https://github.com/folio-org/ui-receiving/tree/v5.0.1) (2024-04-01)